Refactor(docs): Move Storybook to apps
