--- .storybook/assets/stylesheets/index.scss
@@ -1,57 +1,0 @@
-@forward '../../../packages/design-tokens/src/scss';
-
-@forward '../../../packages/web/src/scss/foundation';
-@forward '../../../packages/web/src/scss/helpers';
-
-@forward '../../../packages/web/src/scss/components/Accordion';
-@forward '../../../packages/web/src/scss/components/Alert';
-@forward '../../../packages/web/src/scss/components/Breadcrumbs';
-@forward '../../../packages/web/src/scss/components/Button';
-@forward '../../../packages/web/src/scss/components/Checkbox';
-@forward '../../../packages/web/src/scss/components/Collapse';
-@forward '../../../packages/web/src/scss/components/Container';
-@forward '../../../packages/web/src/scss/components/Dropdown';
-@forward '../../../packages/web/src/scss/components/FieldGroup';
-@forward '../../../packages/web/src/scss/components/FileUploader';
-@forward '../../../packages/web/src/scss/components/Grid';
-@forward '../../../packages/web/src/scss/components/Header';
-@forward '../../../packages/web/src/scss/components/Item';
-@forward '../../../packages/web/src/scss/components/Modal';
-@forward '../../../packages/web/src/scss/components/Pagination';
-@forward '../../../packages/web/src/scss/components/Pill';
-@forward '../../../packages/web/src/scss/components/Radio';
-@forward '../../../packages/web/src/scss/components/ScrollView';
-@forward '../../../packages/web/src/scss/components/Select';
-@forward '../../../packages/web/src/scss/components/Stack';
-@forward '../../../packages/web/src/scss/components/Tabs';
-@forward '../../../packages/web/src/scss/components/Tag';
-@forward '../../../packages/web/src/scss/components/TextArea';
-@forward '../../../packages/web/src/scss/components/TextField';
-@forward '../../../packages/web/src/scss/components/Tooltip';
-
-@forward '../../../packages/web/src/scss/utilities';
-
-.docs-Box {
-    overflow: hidden;
-    min-height: 2rem;
-    padding: 1rem;
-    white-space: nowrap;
-    text-overflow: ellipsis;
-    color: #fff;
-    background: #0b1c21;
-}
-
-.docs-FloatingUI-Wrapper {
-    width: 40rem;
-    max-width: 100%;
-    height: 10rem;
-    overflow: auto;
-}
-
-.docs-FloatingUI-Container {
-    position: relative;
-    width: 100%;
-    height: 20rem;
-    padding-block: 7rem;
-    text-align: center;
-}
