+++ apps/storybook/assets/stylesheets/index.scss
@@ -1,0 +1,31 @@
+@forward '../../../../packages/design-tokens/src/scss';
+
+@forward '../../../../packages/web/src/scss/foundation';
+@forward '../../../../packages/web/src/scss/helpers';
+@forward '../../../../packages/web/src/scss/components';
+@forward '../../../../packages/web/src/scss/utilities';
+
+.docs-Box {
+    overflow: hidden;
+    min-height: 2rem;
+    padding: 1rem;
+    white-space: nowrap;
+    text-overflow: ellipsis;
+    color: #fff;
+    background: #0b1c21;
+}
+
+.docs-FloatingUI-Wrapper {
+    width: 40rem;
+    max-width: 100%;
+    height: 10rem;
+    overflow: auto;
+}
+
+.docs-FloatingUI-Container {
+    position: relative;
+    width: 100%;
+    height: 20rem;
+    padding-block: 7rem;
+    text-align: center;
+}
