+++ apps/storybook/decorators/IconGlobalDecorator.tsx
@@ -2,8 +2,8 @@
 /* eslint-disable import/no-extraneous-dependencies, import/extensions, import/no-unresolved */
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore: No declaration file
-import icons from '../../packages/icons/dist/icons';
 import { IconsProvider } from '@lmc-eu/spirit-web-react/src/context';
+import icons from '../../../packages/icons/dist/icons';
 
 export const IconGlobalDecorator = (Story) => (
   <IconsProvider value={icons}>
