+++ apps/storybook/main.ts
@@ -4,7 +4,7 @@ import { mergeConfig } from 'vite';
 import markdownRawPlugin from 'vite-raw-plugin';
 
 const config: StorybookViteConfig = {
-  stories: ['../packages/**/stories/**/*.mdx', '../packages/**/*.stories.@(ts|tsx)'],
+  stories: ['../../packages/**/*.mdx', '../../packages/**/*.stories.@(ts|tsx)'],
 
   addons: [getAbsolutePath('@storybook/addon-links'), getAbsolutePath('@storybook/addon-essentials')],
 
@@ -29,8 +29,8 @@ const config: StorybookViteConfig = {
         preprocessorOptions: {
           scss: {
             includePaths: [
-              resolve(__dirname, '../node_modules'),
-              resolve(__dirname, '../node_modules/@lmc-eu/spirit-design-tokens/src/scss'),
+              resolve(__dirname, '../../node_modules'),
+              resolve(__dirname, '../../node_modules/@lmc-eu/spirit-design-tokens/src/scss'),
             ],
           },
         },
