+++ apps/storybook/netlify.toml
@@ -2,6 +2,10 @@
 # Settings in the [build] context are global and are applied to
 # all contexts unless otherwise overridden by more specific contexts.
 [build]
+  # Directory to change to before starting a build.
+  # This is where we will look for package.json/.nvmrc/etc.
+  # If not set, defaults to the root directory.
+  base = "apps/storybook/"
 
   # Directory that contains the deploy-ready HTML files and
   # assets generated by the build. This is relative to the base
@@ -9,22 +13,22 @@
   # a base has not been set. This sample publishes the directory
   # located at the absolute path "root/project/build-output"
 
-  publish = ".storybook/build"
+  publish = "build/"
 
   # Default build command.
   command = """\
-    yarn lerna run --scope @lmc-eu/spirit-icons build --skip-nx-cache \
-    && yarn storybook:build \
+    yarn lerna run build --scope @lmc-eu/spirit-icons --skip-nx-cache \
+    && yarn lerna run build --scope @lmc-eu/spirit-storybook --skip-nx-cache \
     """
 
   # Ignore builds on Netlify
-  ignore = "node ./scripts/ignoreNetlifyBuilds.js"
+  ignore = "node ../../scripts/ignoreNetlifyBuilds.js"
 
 [context.deploy-preview]
   # Deploy preview only if there are changes in some directories
   ignore = """\
     git diff --quiet $CACHED_COMMIT_REF $COMMIT_REF \
     ./packages/web-react/src/ \
-    .storybook/ \
+    ./apps/storybook/ \
     ./packages/web-react/docs/stories/ \
     """
