+++ apps/storybook/package.json
@@ -1,0 +1,35 @@
+{
+  "name": "@lmc-eu/spirit-storybook",
+  "version": "0.0.0",
+  "description": "Storybook of the Spirit Design System",
+  "license": "MIT",
+  "private": true,
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/lmc-eu/spirit-design-system.git",
+    "directory": "apps/storybook"
+  },
+  "scripts": {
+    "start": "storybook dev --config-dir ./ --port 6006",
+    "build": "storybook build --config-dir ./ --output-dir ./build"
+  },
+  "dependencies": {},
+  "devDependencies": {
+    "@babel/core": "7.23.7",
+    "@storybook/addon-actions": "7.6.16",
+    "@storybook/addon-essentials": "7.6.16",
+    "@storybook/addon-links": "7.6.16",
+    "@storybook/addons": "7.6.16",
+    "@storybook/blocks": "7.6.16",
+    "@storybook/client-api": "7.6.16",
+    "@storybook/html": "7.6.16",
+    "@storybook/react": "7.6.16",
+    "@storybook/react-vite": "7.6.16",
+    "@storybook/theming": "7.6.16",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "storybook": "7.6.16",
+    "typescript": "^5.4.2",
+    "vite": "4.5.3"
+  }
+}
