+++ lerna.json
@@ -1,6 +1,6 @@
 {
   "$schema": "http://json.schemastore.org/lerna",
-  "packages": ["packages/*", "exporters/*", "apps/demo"],
+  "packages": ["packages/*", "exporters/*", "apps/demo", "apps/storybook"],
   "version": "independent",
   "npmClient": "yarn",
   "useNx": true,
