+++ package.json
@@ -30,8 +30,7 @@
     "format:fix": "prettier --write ./",
     "format:fix:changelog": "prettier --write ./**/CHANGELOG.md",
     "prepare": "is-ci || husky install",
-    "storybook": "storybook dev -p 6006",
-    "storybook:build": "storybook build -o ./.storybook/build",
+    "storybook": "lerna run start --scope=@lmc-eu/spirit-storybook",
     "test": "yarn packages:test",
     "test:e2e": "yarn playwright test",
     "test:e2e:update": "yarn playwright test --update-snapshots",
@@ -66,16 +65,6 @@
     "@rollup/plugin-babel": "6.0.4",
     "@rollup/plugin-node-resolve": "15.2.3",
     "@rollup/plugin-replace": "5.0.5",
-    "@storybook/addon-actions": "7.6.16",
-    "@storybook/addon-essentials": "7.6.16",
-    "@storybook/addon-links": "7.6.16",
-    "@storybook/addons": "7.6.16",
-    "@storybook/blocks": "7.6.16",
-    "@storybook/client-api": "7.6.16",
-    "@storybook/html": "7.6.16",
-    "@storybook/react": "7.6.16",
-    "@storybook/react-vite": "7.6.16",
-    "@storybook/theming": "7.6.16",
     "@svgr/webpack": "8.1.0",
     "babel-loader": "9.1.3",
     "core-js": "3.36.1",
@@ -99,7 +88,6 @@
     "react-syntax-highlighter": "15.5.0",
     "react-textarea-autosize": "8.5.3",
     "sass-loader": "13.3.3",
-    "storybook": "7.6.16",
     "style-loader": "3.3.4",
     "typescript": "4.7.4",
     "vite-raw-plugin": "1.0.2",
